astronomer/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture\#provider-packages

# Your existing dependencies
pandas==1.5.3
apache-airflow-providers-mysql
google-api-python-client
Unidecode
openpyxl
apache-airflow-providers-celery>=3.3.0
apache-airflow-providers-postgres
loadenv==0.1.1
boto3==1.35.0

# Additional providers for your use case
apache-airflow-providers-snowflake
apache-airflow-providers-amazon
snowflake-connector-python
python-dotenv
